// src/entity/permissions.enum.ts
export enum PERMISSIONS {
  //ADMIN
  INVITE_CLIENT = 'INVITE-CLIENT',
  INVITE_TEAM_MEMBER = 'INVITE-TEAM-MEMBER',
  VIEW_CLIENTS = 'VIEW-CLIENTS',
  VIEW_OWN_TEAM_MEMBERS = 'VIEW-OWN-TEAM-MEMBERS',
  VIEW_TEAM_MEMBER_DETAIL = 'VIEW-TEAM-MEMBER-DETAIL',
  UPDATE_TEAM_MEMBER = 'UPDATE-TEAM-MEMBER',
  VIEW_FILES = 'VIEW-FILES',

  //BIOMARKER
  VIEW_BIOMARKERS = 'VIEW-BIOMARKERS',
  CHECK_BIOMARKER_RESULT = 'CHECK-BIOMARKER-RESULT',
  ADD_BIOMARKER = 'ADD-BIOMARKER',
  ADD_LAB_BIOMARKER = 'ADD-LAB-BIOMARKER',
  EDIT_BIOMARKER = 'EDIT-BIOMARKER',
  EDIT_LAB_BIOMARKER = 'EDIT-LAB-BIOMARKER',
  SAVE_ANALYSIS_TAG_INTERPRETATION = 'SAVE-ANALYSIS-TAG-INTERPRETATION',
  VIEW_ANALYSIS_TAG_INTERPRETATION = 'VIEW-ANALYSIS-TAG-INTERPRETATION',

  //COMPONENTS
  VIEW_COMPONENTS = 'VIEW-COMPONENTS',
  ADD_COMPONENT = 'ADD-COMPONENT',

  //LAB TEST
  VIEW_LAB_TESTS = 'VIEW-LAB-TESTS',

  //LAB REPORT
  VIEW_ALL_LAB_REPORTS = 'VIEW-ALL-LAB-REPORTS',
  VIEW_LAB_REPORT = 'VIEW-LAB-REPORT',
  VIEW_REDACTED_LAB_REPORTS = 'VIEW-REDACTED-LAB-REPORTS',
  SUBMIT_LAB_REPORT_DATA = 'SUBMIT-LAB-REPORT-DATA',
  PERSIST_LABORATORY = 'PERSIST-LABORATORY',
  VIEW_REFERRAL_REPORTS = 'VIEW-REFERRAL-REPORTS',
  REFER_LAB_REPORT = 'REFER-LAB-REPORT',
  RESOLVE_REFERRAL_REPORT = 'RESOLVE-REFERRAL-REPORT',
  REDACT_LAB_REPORTS = 'REDACT-LAB-REPORTS',

  //LABORATORY
  VIEW_LABORATORY = 'VIEW-LABORATORY',
  ADD_LABORATORY = 'ADD-LABORATORY',
  EDIT_LABORATORY = 'EDIT-LABORATORY',

  //LEGENDS
  VIEW_LEGENDS = 'VIEW-LEGENDS',

  //UOFM
  VIEW_UNIT_OF_MEASUREMENT = 'VIEW-UNIT-OF-MEASUREMENT',
  ADD_UNIT_OF_MEASUREMENT = 'ADD-UNIT-OF-MEASUREMENT',
  EDIT_UNIT_OF_MEASUREMENT = 'EDIT-UNIT-OF-MEASUREMENT',

  //INTERPRETATION
  ADD_INTERPRETATION = 'ADD-INTERPRETATION',
  EDIT_INTERPRETATION = 'EDIT-INTERPRETATION',
  VIEW_INTERPRETATION = 'VIEW-INTERPRETATION',

  //FILE UPLOAD
  FILE_UPLOAD = 'FILE-UPLOAD',

  //PATHWAY
  VIEW_PATHWAYS = 'VIEW-PATHWAYS',

  //REACTIONS
  VIEW_REACTIONS = 'VIEW-REACTIONS',
  EDIT_REACTIONS = 'EDIT-REACTIONS',

  //RLTNSHIPS
  VIEW_RELATIONSHIPS = 'VIEW-RELATIONSHIPS',
  EDIT_RELATIONSHIP = 'EDIT-RELATIONSHIP',

  //RLTNSHIP PATTERN
  MASTER_BUILD = 'MASTER-BUILD',
  BUILD_RELATIONSHIP = 'BUILD-RELATIONSHIP',
  VIEW_MASTER_BUILD_STATUS = 'VIEW-MASTER-BUILD-STATUS',

  //STATIC DATA
  VIEW_DASHBOARD_DATA = 'VIEW-DASHBOARD-DATA',
  VIEW_BIOMARKERS_STAT = 'VIEW-BIOMARKERS-STAT',
  VIEW_SMPDB_COMPONENTS_STAT = 'VIEW-SMPDB-COMPONENTS-STAT',
  VIEW_BIOMARKERS_SUMMARY = 'VIEW-BIOMARKERS-SUMMARY',

  //USER
  DELETE_USER = 'DELETE-USER',
  VIEW_USERS = 'VIEW-USERS',

  //ROLES
  EDIT_ROLE = 'EDIT-ROLE',
  VIEW_ROLES = 'VIEW-ROLES',
  VIEW_PERMISSIONS = 'VIEW-PERMISSIONS',
  ASSIGN_ROLE = 'ASSIGN-ROLE',
  EDIT_PERMISSION = 'EDIT-PERMISSION',

  //HELP DESK
  VIEW_HELP_DESK_QUERY = 'VIEW-HELP-DESK-QUERY',
  EDIT_HELP_DESK_QUERY = 'EDIT-HELP-DESK-QUERY',

  //WAIT-LISTED-USER
  VIEW_WAITING_LIST_USERS = 'VIEW-WAITING-LIST-USERS',
  UPDATE_WAITING_LIST_USERS = 'UPDATE-WAITING-LIST-USERS',

  //PUBLIC CONTENT
  EDIT_PUBLIC_CONTENT = 'EDIT-PUBLIC-CONTENT',

  //PRACTITIONER
  VIEW_OWN_INFO = 'VIEW-OWN-INFO',
  UPDATE_ONW_INFO = 'UPDATE-ONW-INFO',
  ADD_CLIENT = 'ADD-CLIENT',
  EDIT_CLIENT = 'EDIT-CLIENT',
  UPLOAD_REPORT = 'UPLOAD-REPORT',
  UPDATE_REPORT = 'UPDATE-REPORT',
  VIEW_CLIENT_LAB_REPORT = 'VIEW-CLIENT-LAB-REPORT',
  DELETE_REPORT = 'DELETE-LAB-REPORT',
  VIEW_CLIENT_REPORT_ANALYSIS = 'VIEW-CLIENT-REPORT-ANALYSIS',
  REQUEST_CLIENT_REPORT_ANALYSIS = 'REQUEST-CLIENT-REPORT-ANALYSIS',
  RECOVER_ACCOUNT = 'RECOVER-ACCOUNT',

  //OWNER
  ADD_DATA_ENTRY_USER = 'ADD-DATA-ENTRY-USER',
  UPDATE_DATA_ENTRY_USER_INFO = 'UPDATE-DATA-ENTRY-USER-INFO',
  VIEW_DATA_ENTRY_USER = 'VIEW-DATA-ENTRY-USER',

  // AUDIT LOGS
  VIEW_OWNER_AUDIT_LOGS = 'VIEW-OWNER-AUDIT-LOGS',
  VIEW_PRACTITIONER_AUDIT_LOGS = 'VIEW-PRACTITIONER-AUDIT-LOGS',
  DOWNLOAD_AUDIT_LOGS = 'DOWNLOAD-AUDIT-LOGS',

  //CONTACT US
  VIEW_CONTACT = 'VIEW-CONTACT',

  //SUBSCRIPTION
  VIEW_STRIPE_PRODUCTS = 'VIEW-STRIPE-PRODUCTS',
  CREATE_CHECKOUT_SESSION = 'CREATE-CHECKOUT-SESSION',
  CANCEL_SUBSCRIPTION = 'CANCEL-SUBSCRIPTION',
  UPDATE_SUBSCRIPTION = 'UPDATE-SUBSCRIPTION',
  VIEW_PAYMENT_HISTORY = 'VIEW-PAYMENT-HISTORY',
  VIEW_BILLING_INFO = 'VIEW-BILLING-INFO',
}